pyproject.toml
[tool.poetry]
name = "homework-27"
version = "0.1.0"
description = ""
authors = ["snabi <s.nabievaaa@gmail.com>"]
readme = "README.md"
packages = [{include = "homework_27"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
pillow = "^9.2.0"
psycopg2 = "^2.9.4"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
